feat: organize folder structure
