+++ auth/internal/models/templatedata.go
@@ -1,0 +1,7 @@
+package models
+
+type TemplateData struct {
+	Data      map[string]interface{}
+	CSRFToken string
+	Error     string
+}
